refactor(widgets): 将模型选择弹窗重构为底部卡片形式
- 移除独立的 ModelSelectorPopup 弹窗组件
- 新增 ModelSelectorCardContent 作为底部卡片内容
- 修改主窗口中的模型选择逻辑，从弹窗切换到底部卡片
- 更新模型选择按钮点击事件，改为切换卡片显示
- 调整服务商编辑卡片高度从380px到300px
- 优化卡片注册和显示管理逻辑
- 保持模型选择功能完整性和用户体验一致性
